# ubkg-api-spec.yaml
openapi: 3.0.3
info:
  title: UBKG API
  description: The UBKG API provides REST endpoints that return results from queries against an instance of the [Unified Biomedical Knowledge Graph (UBKG)](https://ubkg.docs.xconsortia.org/). To prevent exceeding service timeouts or payload limits, some endpoints are throttled in both execution time and in the size of its response payload--e.g., to a maximum execution time of 15 seconds and a maximum payload of 9 MB.
  version: "2.0"
  contact:
    name: GitHub repository
    url: https://github.com/x-atlas-consortia/ubkg-api
servers:
  - url: https://ontology.api.hubmapconsortium.org/
    description: Production server
  - url: https://ontology-api.dev.hubmapconsortium.org/
    description: Development server
paths:
...
  /semantics/semantic_types:
    get:
      operationId: semantics_semantic_types_get
      summary: Returns the complete set of Semantic Type nodes, subject to constraints specified in parameters.
      parameters:
        - name: skip
          in: query
          required: false
          description: the number of Semantic Types to skip in the returned set. The value must be non-negative. The default value is 0.
          schema:
            type: string
            example: 0
        - name: limit
          in: query
          required: false
          description: the maximum number of Semantic Types to return. The value must be non-negative. The default value is the maximum number of rows specified by configuration.
          schema:
            type: string
            example: 10
      responses:
        '200':
          description: Returns the set of Semantic Type nodes, subject to constaints specified in the parameters.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SemanticTypes'
        '400':
          description: invalid parameter name; non-numeric parameter value; negative value
        '404':
          description: No subtypes of the specified Semantic Type.
        '5XX':
          description: Unknown error
  /semantics/{identifier}/semantic_types:
    get:
      operationId: semantics_identifier_semantic_types_get
      summary: Returns the Semantic Type node that matches the identifier, subject to constraints specified in parameters.
      parameters:
        - name: identifier
          in: path
          required: true
          description: The identifier for a Semantic Type node. An identifier may be either a URL-encoded string corresponding to either the name of the Semantic Type (e.g., Anatomical%20Structure, with %20 representing the space character) or the Type Unique Identifier (TUI) of the Semantic Type (e.g., T017). (%20 = space; %2C = comma; %3A = colon)
          schema:
            type: string
            example: T017
        - name: skip
          in: query
          required: false
          description: the number of paths to skip in the returned set. The value must be non-negative. The default value is 0.
          schema:
            type: string
            example: 0
        - name: limit
          in: query
          required: false
          description: the maximum number of Semantic Types to return. The value must be non-negative. The default value is the maximum number of rows specified by configuration.
          schema:
            type: string
            example: 10
      responses:
        '200':
          description: Returns the set of Semantic Type nodes, subject to constaints specified in the parameters.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SemanticTypes'
        '400':
          description: invalid parameter name; non-numeric parameter value; negative value
        '404':
          description: No subtypes of the specified Semantic Type.
        '5XX':
          description: Unknown error
  /semantics/{identifier}/subtypes:
    get:
      operationId: semantics_identifier_semantic_subtypes_get
      summary: Returns the set of Semantic Type nodes that are subtypes (have ISA_STY relationship with) the specified Semantic Type, subject to constraints specified in parameters.
      parameters:
        - name: identifier
          in: path
          required: true
          description: The identifier for a Semantic Type node. An identifier may be either a URL-encoded string corresponding to either the name of the Semantic Type (e.g., Anatomical%20Structure, with %20 representing the space character) or the Type Unique Identifier (TUI) of the Semantic Type (e.g., T017). (%20 = space; %2C = comma; %3A = colon)
          schema:
            type: string
            example: Anatomical%20Structure
        - name: skip
          in: query
          required: false
          description: the number of semantic sub-types to skip in the returned set. The value must be non-negative. The default value is 0.
          schema:
            type: string
            example: 0
        - name: limit
          in: query
          required: false
          description: the maximum number of semantic sub-types to return. The value must be non-negative. The default value is the maximum number of rows specified by configuration.
          schema:
            type: string
            example: 10
      responses:
        '200':
          description: Returns the set of Semantic Type nodes that are subtypes of the specified Semantic Type node.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SemanticSubTypes'
        '400':
          description: invalid parameter name; non-numeric parameter value; negative value
        '404':
          description: No subtypes of the specified Semantic Type.
        '5XX':
          description: Unknown error
...
components:
  schemas:
...
    SemanticTypes: # Schema name
      type: object
      description: Array of Semantic Type nodes
      properties:
        position:
          type: integer
          description: absolute position of the Semantic Type in the list of all possible values in the response. Invariant relative to skip--e.g., if skip = 5, the position of the first element in the list will be 6.
          example: 1
        semantic_types:
          type: array
          items:
            type: object
            properties:
              semantic_type:
                type: object
                description: A Semantic Type node
                properties:
                  def:
                    type: string
                    description: the definition for the Semantic Type
                    example: An anatomical structure that exists...
                  stn:
                    type: string
                    description: the Semantic Type's identifier in the Semantic Type Network
                    example: A1.2.1
                  sty:
                    type: string
                    description: the Semantic Type name
                    example: Embryonic Structure
                  tui:
                    type: string
                    description: the Type Unique Identifier (TUI) for the Semantic Type node
                    example: T018
    SemanticSubTypes: # Schema name
      type: object
      description: Array of Semantic Type nodes that are sub-types of a Semantic Type
      properties:
        position:
          type: integer
          description: absolute position of the Semantic Type in the list of all possible values in the response. Invariant relative to skip--e.g., if skip = 5, the position of the first element in the list will be 6.
          example: 1
        semantic_sub_types:
          type: array
          items:
            type: object
            properties:
              semantic_type:
                type: object
                description: A Semantic Type node that is a sub-type of a specified Semantic Type node
                properties:
                  def:
                    type: string
                    description: the definition for the Semantic Type
                    example: An anatomical structure that exists...
                  stn:
                    type: string
                    description: the Semantic Type's identifier in the Semantic Type Network
                    example: A1.2.1
                  sty:
                    type: string
                    description: the Semantic Type name
                    example: Embryonic Structure
                  tui:
                    type: string
                    description: the Type Unique Identifier (TUI) for the Semantic Type node
                    example: T018
...